backend/requirements.txt
# FastAPI Backend Requirements
# Python 3.10+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Machine Learning
torch==2.1.2
torchvision==0.16.2
segmentation-models-pytorch==0.3.3
timm==0.9.12

# Image Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3

# Utilities
python-dotenv==1.0.0

# Albumentation for compatibility
albumentations==1.4.0

